+++ newrelic-api/src/main/java/com/ocado/panda/newrelic/api/internal/NewRelicInternalApi.java
@@ -8,6 +8,7 @@ import com.ocado.panda.newrelic.api.ApplicationsApi;
 import com.ocado.panda.newrelic.api.DeploymentsApi;
 import com.ocado.panda.newrelic.api.KeyTransactionsApi;
 import com.ocado.panda.newrelic.api.ServersApi;
+import com.ocado.panda.newrelic.api.UsersApi;
 import com.ocado.panda.newrelic.api.internal.client.NewRelicClient;
 import lombok.Getter;
 
@@ -35,6 +36,8 @@ public class NewRelicInternalApi {
 
     private final ServersApi serversApi;
 
+    private final UsersApi usersApi;
+
     /**
      * NewRelic API constructor.
      *
@@ -51,5 +54,6 @@ public class NewRelicInternalApi {
         keyTransactionsApi = new DefaultKeyTransactionsApi(client);
         deploymentsApi = new DefaultDeploymentsApi(client);
         serversApi = new DefaultServersApi(client);
+        usersApi = new DefaultUsersApi(client);
     }
 }